Moved replace function into separate header
