+++ CMakeLists.txt
@@ -25,9 +25,9 @@ include(cmake/Git.cmake)
 ### SETTINGS ###
 
 # add all headers (.h, .hpp) to this
-# set(PRJ_HEADERS)
+set(PRJ_HEADERS src/replace.h)
 # add all source files (.cpp) to this, except the one with main()
-# set(PRJ_SOURCES src/main.cpp)
+set(PRJ_SOURCES src/replace.cpp)
 # set the source file containing main()
 set(PRJ_MAIN src/main.cpp)
 # set include paths not part of libraries
